requirements.txt
alembic==1.5.5
blinker==1.4
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyJWT==2.3.0
pymongo==4.0.2
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3